refactor: move CreateAnalysis component to dedicated subdirectory
