+++ src/ravioli/frontend/src/main.ts
@@ -4,7 +4,7 @@ import { api } from './services/api';
 import { renderSidebar, updateSidebarUI } from './components/Sidebar';
 import { renderAnalysis, updateAnalysisUI } from './components/analysis/AnalysisShell';
 import { renderInsights } from './components/Insights';
-import { renderCreateAnalysis } from './components/analysis/CreateAnalysis';
+import { renderCreateAnalysis } from './components/analysis/create-analysis/CreateAnalysis';
 import { renderKnowledge } from './components/Knowledge';
 import { renderData } from './components/Data';
 import { renderSettings } from './components/Settings';
